requirements.txt
Flask==3.0.3
groq==0.8.0
httpx==0.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
